Greatly simplify the creation and deployment of locally-built Docker images.
This change removes the per-developer images uploaded to gcr.io, and in their
place adds an install.sh script which allows installing a local Docker image to
an arbitrary GCP project.

In order to facilitate the creation of the new install.sh script, I had to
change the layering of how the connection to the Cloud SQL instance is
configured.

Previously, the user/developer had to manually setup a username/password for
the Cloud SQL instance, and then inject those into the Docker container via
environment variables. Automating that process proved problematic, so instead
the user will only inject the name of the Cloud SQL instance, and the Docker
container will set the root password on that instance using the gcloud tool.

This has the downside of tying our Docker image to Cloud SQL, but in exchange
for that we get to make our installation process more automated and repeatable.